feat(api): deprecate nvim_out/err_write(ln)
