+++ runtime/doc/deprecated.txt
@@ -18,6 +18,9 @@ DEPRECATED IN 0.11					*deprecated-0.11*
 API
 • nvim_subscribe()	Plugins must maintain their own "multicast" channels list.
 • nvim_unsubscribe()	Plugins must maintain their own "multicast" channels list.
+• nvim_out_write()	Use |nvim_echo()|.
+• nvim_err_write()	Use |nvim_echo()| with `{err=true}`.
+• nvim_err_writeln()	Use |nvim_echo()| with `{err=true}`.
 
 DIAGNOSTICS
 • *vim.diagnostic.goto_next()*	Use |vim.diagnostic.jump()| with `{count=1, float=true}` instead.
